requirements.txt
python-telegram-bot>=20.5
redis>=4.5.4
aiosqlite>=0.18.0
